# packages/contracts/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "oMLX management API",
    "version": "1"
  },
  "paths": {
...
    "/management/v1/models/{model_id}/load": {
      "post": {
        "tags": [
          "management"
        ],
        "summary": "Load Model",
        "operationId": "load_model_management_v1_models__model_id__load_post",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelOperationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/management/v1/models/{model_id}/unload": {
      "post": {
        "tags": [
          "management"
        ],
        "summary": "Unload Model",
        "operationId": "unload_model_management_v1_models__model_id__unload_post",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelOperationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/management/v1/settings": {
      "get": {
        "tags": [
          "management"
        ],
        "summary": "Get Global Settings",
        "operationId": "get_global_settings_management_v1_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalSettingsView"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "management"
        ],
        "summary": "Update Global Settings",
        "operationId": "update_global_settings_management_v1_settings_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlobalSettingsPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalSettingsUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/management/v1/models/{model_id}/settings": {
      "get": {
        "tags": [
          "management"
        ],
        "summary": "Get Model Settings",
        "operationId": "get_model_settings_management_v1_models__model_id__settings_get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelSettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "management"
        ],
        "summary": "Update Model Settings",
        "operationId": "update_model_settings_management_v1_models__model_id__settings_patch",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelSettingsPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelSettingsUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/management/v1/models/{model_id}/profiles": {
      "get": {
        "tags": [
          "management"
        ],
        "summary": "List Profiles",
        "operationId": "list_profiles_management_v1_models__model_id__profiles_get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfilesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "management"
        ],
        "summary": "Create Profile",
        "operationId": "create_profile_management_v1_models__model_id__profiles_post",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/management/v1/models/{model_id}/profiles/{name}": {
      "put": {
        "tags": [
          "management"
        ],
        "summary": "Update Profile",
        "operationId": "update_profile_management_v1_models__model_id__profiles__name__put",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "management"
        ],
        "summary": "Delete Profile",
        "operationId": "delete_profile_management_v1_models__model_id__profiles__name__delete",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/management/v1/models/{model_id}/profiles/{name}/apply": {
      "post": {
        "tags": [
          "management"
        ],
        "summary": "Apply Profile",
        "operationId": "apply_profile_management_v1_models__model_id__profiles__name__apply_post",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelSettingsUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/management/v1/stats": {
      "get": {
        "tags": [
          "management"
        ],
        "summary": "Get Stats",
        "operationId": "get_stats_management_v1_stats_get",
        "parameters": [
          {
            "name": "model_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Model Id"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(session|alltime)$",
              "default": "session",
              "title": "Scope"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/management/v1/cache/{kind}/clear": {
      "post": {
        "tags": [
          "management"
        ],
        "summary": "Clear Cache",
        "operationId": "clear_cache_management_v1_cache__kind__clear_post",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CacheClearResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/management/v1/diffusion/calibrations": {
      "post": {
        "tags": [
          "management"
        ],
        "summary": "Start Diffusion Calibration",
        "operationId": "start_diffusion_calibration_management_v1_diffusion_calibrations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiffusionCalibrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiffusionJobView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/management/v1/diffusion/quantizations": {
      "post": {
        "tags": [
          "management"
        ],
        "summary": "Start Diffusion Quantization",
        "operationId": "start_diffusion_quantization_management_v1_diffusion_quantizations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiffusionQuantizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiffusionJobView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/management/v1/diffusion/jobs/{job_id}": {
      "get": {
        "tags": [
          "management"
        ],
        "summary": "Get Diffusion Job",
        "operationId": "get_diffusion_job_management_v1_diffusion_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiffusionJobView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/management/v1/diffusion/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "management"
        ],
        "summary": "Cancel Diffusion Job",
        "operationId": "cancel_diffusion_job_management_v1_diffusion_jobs__job_id__cancel_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiffusionJobView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GlobalSettingsPatch": {
        "properties": {
          "max_context_window": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Context Window"
          },
          "max_context_window_policy": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Context Window Policy"
          },
          "max_tokens": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "top_p": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Top P"
          },
          "top_k": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Top K"
          },
          "repetition_penalty": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Repetition Penalty"
          },
          "max_concurrent_requests": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Concurrent Requests"
          },
          "embedding_batch_size": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Batch Size"
          },
          "chunked_prefill": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunked Prefill"
          },
          "prefill_priority": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "context",
                  "speed"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Prefill Priority"
          },
          "decode_fairness": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decode Fairness"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "GlobalSettingsPatch"
      },
      "GlobalSettingsUpdateResponse": {
        "properties": {
          "sampling": {
            "$ref": "#/components/schemas/SamplingSettingsView"
          },
          "scheduler": {
            "$ref": "#/components/schemas/SchedulerSettingsView"
          },
          "requires_restart": {
            "type": "boolean",
            "title": "Requires Restart"
          }
        },
        "type": "object",
        "required": [
          "sampling",
          "scheduler",
          "requires_restart"
        ],
        "title": "GlobalSettingsUpdateResponse"
      },
      "GlobalSettingsView": {
        "properties": {
          "sampling": {
            "$ref": "#/components/schemas/SamplingSettingsView"
          },
          "scheduler": {
            "$ref": "#/components/schemas/SchedulerSettingsView"
          }
        },
        "type": "object",
        "required": [
          "sampling",
          "scheduler"
        ],
        "title": "GlobalSettingsView"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelOperationResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "unloading"
            ],
            "title": "Status"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "status",
          "model_id"
        ],
        "title": "ModelOperationResponse"
      },
...
      "ModelSettingsResponse": {
        "properties": {
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "settings": {
            "$ref": "#/components/schemas/ModelSettingsView"
          }
        },
        "type": "object",
        "required": [
          "model_id",
          "settings"
        ],
        "title": "ModelSettingsResponse"
      },
      "ModelSettingsUpdateResponse": {
        "properties": {
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "settings": {
            "$ref": "#/components/schemas/ModelSettingsView"
          },
          "requires_reload": {
            "type": "boolean",
            "title": "Requires Reload"
          },
          "auto_unloaded": {
            "type": "boolean",
            "title": "Auto Unloaded"
          },
          "auto_reloaded": {
            "type": "boolean",
            "title": "Auto Reloaded"
          },
          "reload_deferred": {
            "type": "boolean",
            "title": "Reload Deferred"
          },
          "reload_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reload Error"
          }
        },
        "type": "object",
        "required": [
          "model_id",
          "settings",
          "requires_reload",
          "auto_unloaded",
          "auto_reloaded",
          "reload_deferred",
          "reload_error"
        ],
        "title": "ModelSettingsUpdateResponse"
      },
...
      "ProfilesResponse": {
        "properties": {
          "profiles": {
            "items": {
              "$ref": "#/components/schemas/ProfileView"
            },
            "type": "array",
            "title": "Profiles"
          }
        },
        "type": "object",
        "required": [
          "profiles"
        ],
        "title": "ProfilesResponse"
      },
...
    }
  }
}